# internal/controllers/pengajuan_controller.go
package controllers

import (
	"strconv"

	"rires-be/internal/dto/request"
	"rires-be/internal/dto/response"
	"rires-be/pkg/services"
	"rires-be/pkg/utils"

	"github.com/go-playground/validator/v10"
	"github.com/gofiber/fiber/v2"
)

// PengajuanController handles mahasiswa PKM submission endpoints
type PengajuanController struct {
	service   *services.PengajuanService
	validator *validator.Validate
}

// NewPengajuanController creates a new controller instance
func NewPengajuanController() *PengajuanController {
	return &PengajuanController{
		service:   services.NewPengajuanService(),
		validator: validator.New(),
	}
}

// CreateJudulPKM godoc
// @Summary Create PKM Title Submission
// @Description Mahasiswa (ketua) creates new PKM title submission with team members
// @Tags Mahasiswa - Pengajuan PKM
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer token"
// @Param body body request.CreatePengajuanRequest true "Pengajuan data"
// @Success 201 {object} response.APIResponse{data=response.PengajuanResponse}
// @Failure 400 {object} response.APIResponse
// @Failure 401 {object} response.APIResponse
// @Failure 500 {object} response.APIResponse
// @Security BearerAuth
// @Router /pengajuan/judul [post]
func (ctrl *PengajuanController) CreateJudulPKM(c *fiber.Ctx) error {
	// 1. Parse request body
	var req request.CreatePengajuanRequest
	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Invalid request body",
			err.Error(),
		))
	}

	// 2. Validate request
	if err := ctrl.validator.Struct(req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Validation failed",
			ctrl.formatValidationErrors(err),
		))
	}

	// 3. Custom validation
	if err := req.Validate(); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Validation failed",
			err.Error(),
		))
	}

	// 4. Get authenticated user (NIM ketua)
	nimKetua := utils.GetCurrentUsername(c)

	// 5. Call service
	result, err := ctrl.service.CreateJudulPKM(&req, nimKetua)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Failed to create pengajuan",
			err.Error(),
		))
	}

	// 6. Return success response
	return c.Status(fiber.StatusCreated).JSON(response.SuccessResponse(
		"Pengajuan berhasil dibuat",
		result,
	))
}

// GetMySubmissions godoc
// @Summary Get My Submissions
// @Description Get all submissions created by authenticated mahasiswa
// @Tags Mahasiswa - Pengajuan PKM
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer token"
// @Param status query string false "Filter by status (all/pending/acc/revisi/tolak)"
// @Success 200 {object} response.APIResponse{data=[]response.PengajuanListResponse}
// @Failure 401 {object} response.APIResponse
// @Failure 500 {object} response.APIResponse
// @Security BearerAuth
// @Router /pengajuan/my-submissions [get]
func (ctrl *PengajuanController) GetMySubmissions(c *fiber.Ctx) error {
	// 1. Get authenticated user NIM
	nimKetua := utils.GetCurrentUsername(c)

	// 2. Get filter from query params
	statusFilter := c.Query("status", "all")

	// 3. Call service
	result, err := ctrl.service.GetMySubmissions(nimKetua, statusFilter)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(response.ErrorResponse(
			"Failed to get submissions",
			err.Error(),
		))
	}

	// 4. Return success
	return c.JSON(response.SuccessResponse(
		"Submissions retrieved successfully",
		result,
	))
}

// GetPengajuanDetail godoc
// @Summary Get Pengajuan Detail
// @Description Get detailed information of a pengajuan (only if user is part of the team)
// @Tags Mahasiswa - Pengajuan PKM
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer token"
// @Param id path int true "Pengajuan ID"
// @Success 200 {object} response.APIResponse{data=response.PengajuanResponse}
// @Failure 400 {object} response.APIResponse
// @Failure 401 {object} response.APIResponse
// @Failure 403 {object} response.APIResponse
// @Failure 404 {object} response.APIResponse
// @Failure 500 {object} response.APIResponse
// @Security BearerAuth
// @Router /pengajuan/{id} [get]
func (ctrl *PengajuanController) GetPengajuanDetail(c *fiber.Ctx) error {
	// 1. Parse ID from URL
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Invalid pengajuan ID",
			err.Error(),
		))
	}

	// 2. Get detail from service
	result, err := ctrl.service.GetPengajuanDetail(id)
	if err != nil {
		return c.Status(fiber.StatusNotFound).JSON(response.ErrorResponse(
			"Pengajuan not found",
			err.Error(),
		))
	}

	// 3. Check authorization - only team members can view
	// TODO: Verify user is part of the team
	// nimUser := utils.GetCurrentUsername(c)
	// if !isTeamMember(result, nimUser) {
	//     return c.Status(403).JSON(...)
	// }

	// 4. Return success
	return c.JSON(response.SuccessResponse(
		"Pengajuan detail",
		result,
	))
}

// UpdateJudul godoc
// @Summary Update/Revise PKM Title
// @Description Ketua revises PKM title (only allowed when status = REVISI)
// @Tags Mahasiswa - Pengajuan PKM
// @Accept json
// @Produce json
// @Param Authorization header string true "Bearer token"
// @Param id path int true "Pengajuan ID"
// @Param body body request.UpdateJudulRequest true "Updated title data"
// @Success 200 {object} response.APIResponse{data=response.PengajuanResponse}
// @Failure 400 {object} response.APIResponse
// @Failure 401 {object} response.APIResponse
// @Failure 403 {object} response.APIResponse
// @Failure 404 {object} response.APIResponse
// @Failure 500 {object} response.APIResponse
// @Security BearerAuth
// @Router /pengajuan/judul/{id} [put]
func (ctrl *PengajuanController) UpdateJudul(c *fiber.Ctx) error {
	// 1. Parse ID from URL
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Invalid pengajuan ID",
			err.Error(),
		))
	}

	// 2. Parse request body
	var req request.UpdateJudulRequest
	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Invalid request body",
			err.Error(),
		))
	}

	// 3. Validate request
	if err := ctrl.validator.Struct(req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Validation failed",
			ctrl.formatValidationErrors(err),
		))
	}

	// 4. Get authenticated user
	nimKetua := utils.GetCurrentUsername(c)

	// 5. Call service
	result, err := ctrl.service.UpdateJudul(id, &req, nimKetua)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Failed to update judul",
			err.Error(),
		))
	}

	// 6. Return success
	return c.JSON(response.SuccessResponse(
		"Judul berhasil direvisi",
		result,
	))
}

// UploadProposal godoc
// @Summary Upload Proposal File
// @Description Ketua uploads proposal PDF/DOC/DOCX (only allowed when status_judul = ACC)
// @Tags Mahasiswa - Pengajuan PKM
// @Accept multipart/form-data
// @Produce json
// @Param Authorization header string true "Bearer token"
// @Param id path int true "Pengajuan ID"
// @Param file formData file true "Proposal file (PDF/DOC/DOCX, max 2.5MB)"
// @Success 200 {object} response.APIResponse{data=response.PengajuanResponse}
// @Failure 400 {object} response.APIResponse
// @Failure 401 {object} response.APIResponse
// @Failure 403 {object} response.APIResponse
// @Failure 404 {object} response.APIResponse
// @Failure 500 {object} response.APIResponse
// @Security BearerAuth
// @Router /pengajuan/{id}/proposal [post]
func (ctrl *PengajuanController) UploadProposal(c *fiber.Ctx) error {
	// 1. Parse ID from URL
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Invalid pengajuan ID",
			err.Error(),
		))
	}

	// 2. Get file from form
	file, err := c.FormFile("file")
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"File is required",
			err.Error(),
		))
	}

	// 3. Get authenticated user
	nimKetua := utils.GetCurrentUsername(c)

	// 4. Call service
	result, err := ctrl.service.UploadProposal(id, file, nimKetua)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Failed to upload proposal",
			err.Error(),
		))
	}

	// 5. Return success
	return c.JSON(response.SuccessResponse(
		"Proposal berhasil diupload",
		result,
	))
}

// ReviseProposal godoc
// @Summary Revise Proposal File
// @Description Ketua revises proposal by uploading new file (only allowed when status_proposal = REVISI)
// @Tags Mahasiswa - Pengajuan PKM
// @Accept multipart/form-data
// @Produce json
// @Param Authorization header string true "Bearer token"
// @Param id path int true "Pengajuan ID"
// @Param file formData file true "Revised proposal file (PDF/DOC/DOCX, max 2.5MB)"
// @Success 200 {object} response.APIResponse{data=response.PengajuanResponse}
// @Failure 400 {object} response.APIResponse
// @Failure 401 {object} response.APIResponse
// @Failure 403 {object} response.APIResponse
// @Failure 404 {object} response.APIResponse
// @Failure 500 {object} response.APIResponse
// @Security BearerAuth
// @Router /pengajuan/{id}/proposal [put]
func (ctrl *PengajuanController) ReviseProposal(c *fiber.Ctx) error {
	// 1. Parse ID from URL
	id, err := strconv.Atoi(c.Params("id"))
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Invalid pengajuan ID",
			err.Error(),
		))
	}

	// 2. Get file from form
	file, err := c.FormFile("file")
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"File is required",
			err.Error(),
		))
	}

	// 3. Get authenticated user
	nimKetua := utils.GetCurrentUsername(c)

	// 4. Call service
	result, err := ctrl.service.ReviseProposal(id, file, nimKetua)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(response.ErrorResponse(
			"Failed to revise proposal",
			err.Error(),
		))
	}

	// 5. Return success
	return c.JSON(response.SuccessResponse(
		"Proposal berhasil direvisi",
		result,
	))
}

// ========================================
// HELPER FUNCTIONS
// ========================================

// formatValidationErrors formats validator errors to readable format
func (ctrl *PengajuanController) formatValidationErrors(err error) []response.ValidationErrorResponse {
	var errors []response.ValidationErrorResponse

	if validationErrs, ok := err.(validator.ValidationErrors); ok {
		for _, e := range validationErrs {
			errors = append(errors, response.ValidationErrorResponse{
				Field:   e.Field(),
				Message: ctrl.getValidationMessage(e),
			})
		}
	}

	return errors
}

// getValidationMessage returns user-friendly validation message
func (ctrl *PengajuanController) getValidationMessage(e validator.FieldError) string {
	switch e.Tag() {
	case "required":
		return e.Field() + " is required"
	case "min":
		return e.Field() + " must be at least " + e.Param()
	case "max":
		return e.Field() + " must be at most " + e.Param()
	case "oneof":
		return e.Field() + " must be one of: " + e.Param()
	default:
		return e.Field() + " is invalid"
	}
}